pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "trellis"
description = "Trellis"
authors = []
dependencies = [
    "torch==2.4.1",
    "torchvision",
    "xformers==0.0.28",
    "kaolin==0.18.0",
    "spconv-cu118",
    "pipelime-python",
    "easydict",
    "rembg",
    "onnxruntime",
    "transformers",
    "open3d",
    "plyfile",
    "xatlas",
    "pyvista",
    "pymeshfix",
    "igraph",
    "flash-attn==2.7.3+cu118torch2.4",
    "wheel",

    "gsplat==1.5.1+pyt241cu118",
    "eyesplat",
    "eyecan-calibry[refiners,feats]==0.4.0",
    "cutie",
]
dynamic = ["version"]
requires-python = ">=3.11, < 3.12"
readme = "README.md"
license = { text = "MIT" }

[project.optional-dependencies]
dev = ["pytest", "pytest-cov", "hatch", "build"]

[project.scripts]
trellis = "trellis.__init__:app"

[tool.black]
target-version = ['py311']

[tool.pytest.ini_options]
filterwarnings = ["ignore::DeprecationWarning"]
tmp_path_retention_count = 0
tmp_path_retention_policy = "failed"

[tool.hatch.version]
path = "trellis/__init__.py"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build]
include = [
    "trellis/**/*.py",
]
